Changed from 4 bit register to 8 bit register.
notes: 19 external LED's were added to the basys - 3 via the pmod gpio pins. JC, JB, and JA. 3 LEDs are program clock related, the others are two register outputs for visualization.
What needs to be added now is a demultixplexer via a0,a1,a2,a3, 4 new inputs must be created to be AND'd with the load and enable in order to decide which location in memory to store and retrieve from.